LS980: overlay: Add overlay to prevent tearing down CDMA Phone in favor of GSM.
This should fix the situation where voice radio tech reports as LTE but Voice Over LTE (VoLTE) is not supported (Sprint network). By setting it to 6 (1x) it defaults the voice tech to CDMAPhone and avoids the being set to GSMPhone causing voice calls to fail.

Thanks to toastcfh for suggesting a cleaner alternative.

Change-Id: I36748de4e85377e190436ee0d6ace33298c364a0